feat: добавить друзей и лайки

+++ src/main/java/ru/yandex/practicum/filmorate/util/FilmIDGenerator.java
@@ -1,9 +1,9 @@
 package ru.yandex.practicum.filmorate.util;
 
 public class FilmIDGenerator {
-    private static int id = 0;
+    private static long id = 0;
 
-    public static int incrementAndGetFilmId() {
+    public static long incrementAndGetFilmId() {
         return ++id;
     }
